requirements.txt
python-telegram-bot
beautifulsoup4
requests
psycopg2
pytz
emoji